pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "barygraph-kaikki"
version = "0.4.0"
description = "BaryGraph proof-of-concept on the kaikki.org English dictionary corpus"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "pymongo>=4.6",
  "dnspython>=2.4",
  "numpy>=1.26",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "tqdm>=4.66",
  "orjson>=3.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "mypy>=1.8",
  "mongomock>=4.1",
]

[project.scripts]
bary = "scripts._base:main"

[tool.hatch.build.targets.wheel]
packages = ["lib", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
files = ["lib", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
  "integration: requires a running MongoDB (docker compose up)",
]
